+++ src/camera.hpp
@@ -12,9 +12,9 @@ class Camera
   glm::vec3 velocity = glm::vec3(0.0f);
   glm::vec3 position = glm::vec3(0.0f, 0.3f, 0.0f);
   float pitch { 0.0f };
-  float deltaPitch { 0.0f };
+  double deltaPitch { 0.0f };
   float yaw { 0.0f };
-  float deltaYaw { 0.0f };
+  double deltaYaw { 0.0f };
   float moveSpeed { 1.0f };
   float pitchSpeed { 2.0f };
   float yawSpeed{ 2.0f };
